fix some C4244 warnings
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4795 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/DeCSS/udf.cpp
@@ -14,7 +14,7 @@ static bool aspi_GetSectorInfo(const HANDLE hDrive, DWORD* sec_size, DWORD* max_
 	GetFileSizeEx(hDrive, &size);
 
 	*sec_size = 2048;
-	*max_sec = size.QuadPart / *sec_size;
+	*max_sec = (DWORD)(size.QuadPart / *sec_size);
 
 	return true;
 }
